add convertTorchToCcl
